bug: 카카오 로그인 앱으로 로그인 오류 해결 (#43)
