+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/login/entrance/EntranceFragment.kt
@@ -2,12 +2,15 @@ package com.teameetmeet.meetmeet.presentation.login.entrance
 
 import android.os.Bundle
 import android.view.View
-import android.widget.Toast
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.findNavController
+import com.kakao.sdk.auth.model.OAuthToken
+import com.kakao.sdk.common.model.ClientError
+import com.kakao.sdk.common.model.ClientErrorCause
+import com.kakao.sdk.user.UserApiClient
 import com.teameetmeet.meetmeet.R
 import com.teameetmeet.meetmeet.databinding.FragmentEntranceBinding
 import com.teameetmeet.meetmeet.presentation.base.BaseFragment
@@ -27,7 +30,7 @@ class EntranceFragment : BaseFragment<FragmentEntranceBinding>(R.layout.fragment
 
     private fun setClickListener() {
         binding.loginIbKakaoLogin.setOnClickListener {
-            viewModel.loginKakao()
+            loginKakao()
         }
 
         binding.loginTvAppLogin.setOnClickListener {
@@ -40,6 +43,40 @@ class EntranceFragment : BaseFragment<FragmentEntranceBinding>(R.layout.fragment
             findNavController().navigate(
                 EntranceFragmentDirections.actionEntranceFragmentToSignUpFragment()
             )
+        }
+    }
+
+    private val kakaoLoginCallback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
+        if (error != null) {
+            showMessage(R.string.login_kakao_message_kakao_login_fail, error.message.orEmpty())
+        } else if (token != null) {
+            viewModel.loginApp()
+        }
+    }
+
+    private fun loginKakao() {
+        lifecycleScope.launch {
+            if (UserApiClient.instance.isKakaoTalkLoginAvailable(requireContext())) {
+                UserApiClient.instance.loginWithKakaoTalk(requireContext()) { token, error ->
+                    if (error != null) {
+                       showMessage(R.string.login_kakao_message_kakao_login_fail, error.message.orEmpty())
+                        if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
+                            return@loginWithKakaoTalk
+                        }
+                        UserApiClient.instance.loginWithKakaoAccount(
+                            requireContext(),
+                            callback = kakaoLoginCallback
+                        )
+                    } else if (token != null) {
+                        viewModel.loginApp()
+                    }
+                }
+            } else {
+                UserApiClient.instance.loginWithKakaoAccount(
+                    requireContext(),
+                    callback = kakaoLoginCallback
+                )
+            }
         }
     }
